make ready to deploy on vercel: - build.js should do all the building, not server.js (???) - added vercel.json w/ vercel config - seperated out express stuff so it can be run both locally and on vercel
