+++ vercel.json
@@ -1,0 +1,5 @@
+{
+  "version": 2,
+  "cleanUrls": true,
+  "rewrites": [{ "source": "/(.*)", "destination": "/api" }]
+}
